+++ components/SectionTitle.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="mb-[30px] font-jost text-[22px] tracking-[2px]">
+    <div class="mb-[30px] font-jost text-[22px]">
         <slot />
     </div>
 </template>