+++ components/StickySection.vue
@@ -1,0 +1,18 @@
+<template>
+    <section class="sticky top-0" :style="{ width, background }">
+        <slot></slot>
+    </section>
+</template>
+
+<script setup>
+defineProps({
+    width: {
+        type: String,
+        default: 'auto',
+    },
+    background: {
+        type: String,
+        default: 'none',
+    },
+})
+</script>