+++ components/post-content/Content.vue
@@ -1,6 +1,6 @@
 <template>
     <div
-        class="c-content text-[20px] font-jost relative border-b-[1px] border-solid border-[#eee]"
+        class="c-content text-[20px] font-jost relative border-b-[1px] border-solid border-transparent"
         v-html="convertMarkdownToHTML(content)"
     ></div>
 </template>
@@ -19,14 +19,14 @@ import hljs from 'highlight.js'
 import 'highlight.js/styles/atom-one-dark-reasonable.css'
 
 const md = new MarkdownIt({
-  highlight: function (str, lang) {
-    const code = hljs.highlight(str, {
-        language: lang,
-        ignoreIllegals: true,
-    }).value
+    highlight: function (str, lang) {
+        const code = hljs.highlight(str, {
+            language: lang,
+            ignoreIllegals: true,
+        }).value
 
-    return `<pre class="hljs icon-${lang}"><code>${code}</code></pre>`
-  }
+        return `<pre class="hljs icon-${lang}"><code>${code}</code></pre>`
+    }
 })
     .use(highlightLines)
     .use(codetabs)
@@ -301,7 +301,7 @@ defineProps({
 
 .c-content blockquote {
     margin: 0;
-    background-color: #c4c4c41a;
+    background-color: #eee;
     border-radius: 14px;
     font-style: italic;
     margin: 7px;
@@ -322,7 +322,7 @@ defineProps({
         }
     }
 
-    &::before {
+    /* &::before {
         content: '„';
         position: absolute;
         bottom: -40px;
@@ -342,6 +342,6 @@ defineProps({
         font-size: 100px;
         font-family: monospace;
         color: #ececec;
-    }
+    } */
 }
 </style>