+++ layouts/default.vue
@@ -2,31 +2,22 @@
     <Head>
         <Meta name="theme-color" content="#fff" />
     </Head>
-    <div class="wrapper">
-        <div>
-          <baner-left/>
-        </div>
-        <div>
-            <simplebar class="component-border-vertical h-full">
-                <NuxtPage />
-            </simplebar>
-        </div>
-        <div>
-            <simplebar class="component-border-vertical h-full ">
-              <section class="fixed">
-                <OtherPosts v-if="!isHomePage" />
-                <ExternalLinks />
-              </section>
-            </simplebar>
-        </div>
-    </div>
+    
+    <MainContainer>
+        <NuxtPage />
+    </MainContainer>
+
+    <AsideContainer class="component-border-vertical">
+        <BasicSection v-if="!isHomePage">
+            <OtherPosts />
+        </BasicSection>
+        <BasicSection>
+            <ExternalLinks />
+        </BasicSection>
+    </AsideContainer>
 </template>
 
 <script setup>
-import simplebar from "simplebar-vue"
-import "simplebar/dist/simplebar.min.css"
-
 const router = useRouter()
-
 const isHomePage = computed(() => router.currentRoute.value.path === "/")
 </script>