+++ pages/[...page].vue
@@ -4,40 +4,98 @@
     <Meta name="description" :content="data?.post?.teaser" />
   </Head>
 
-  <Header
-    v-if="!pendingPost && data?.post?.id"
-    :image="data?.post?.main_image_url ?? ''"
-    :name="data?.post?.title ?? ''"
-    :timeAgo="data?.post?.modified_at ?? ''"
-    :teaser="data?.post?.teaser ?? ''"
-    :authorName="data?.postAuthor?.display_name ?? ''"
-    :authorPhoto="'https://images.necodeo.com/' + data?.postAuthor?.image_id_avatar ?? ''"
-    :postId="data?.post?.id ?? ''"
-    :averageRating="data?.post?.rating_average ?? 0"
-    :numberOfComments="data?.post?.comments_count"
-  />
-  <HeaderPlaceholder
-    v-else
-    image=""
-    name="Aliquam erat volutpat"
-    timeAgo="2024-01-01 03:37:37"
-    teaser="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec euismod, nisl eget ultricies ultrices, nisl nisl ultricies nisl, nec ultricies"
-    authorName="John Doe"
-    authorPhoto=""
-    postId=""
-    :averageRating="0"
-    :numberOfComments="0"
-  />
+  <SectionWrapper width="var(--desktop-main-content-width)">
+    <template #aside>
+      <StickySection width="340px">
+        <div class="background m-[7px]">
+          <AdPlaceholder />
+        </div>
+      </StickySection>
+    </template>
 
-  <Content v-if="!pendingPost && data?.post?.id" :content="data?.post?.content ?? ''" />
-  <ContentPlaceholder v-else />
+    <!-- <BasicSection width="var(--main-width)" class="not-desktop">
+      TabletAdv
+    </BasicSection> -->
 
-  <div class="component-border-horizontal font-jost p-[30px]">
-    <PostAuthor v-if="!pendingPost && data?.post?.id" :profile="data.postAuthor" />
-    <PostAuthorPlaceholder v-else />
-  </div>
+    <BasicSection width="var(--main-width)" class="component-border-vertical">
+      <div>
+        <Header
+          v-if="!pendingPost && data?.post?.id"
+          :image="data?.post?.main_image_url ?? ''"
+          :name="data?.post?.title ?? ''"
+          :timeAgo="data?.post?.modified_at ?? ''"
+          :teaser="data?.post?.teaser ?? ''"
+          :authorName="data?.postAuthor?.display_name ?? ''"
+          :authorPhoto="'https://images.necodeo.com/' + data?.postAuthor?.image_id_avatar ?? ''"
+          :postId="data?.post?.id ?? ''"
+          :averageRating="data?.post?.rating_average ?? 0"
+          :numberOfComments="data?.post?.comments_count"
+        />
+        <HeaderPlaceholder
+          v-else
+          image=""
+          name="Aliquam erat volutpat"
+          timeAgo="2024-01-01 03:37:37"
+          teaser="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec euismod, nisl eget ultricies ultrices, nisl nisl ultricies nisl, nec ultricies"
+          authorName="John Doe"
+          authorPhoto=""
+          postId=""
+          :averageRating="0"
+          :numberOfComments="0"
+        />
+      </div>
+    </BasicSection>
+  </SectionWrapper>
 
-  <Comments id="comments" :postId="data?.post?.id" />
+  <SectionWrapper width="var(--desktop-main-content-width)">
+    <template #aside>
+      <StickySection width="340px">
+        <div class="m-[7px]">
+          <ContentNav :items="[
+            {
+              title: 'Integer ornare efficitur risus quis placerat',
+              url: '#integer-ornare-efficitur-risus-quis-placerat',
+            },
+            // {
+            //   title: 'Ut mollis commodo mi, sed rutrum ipsum fringilla at',
+            //   url: '#ut-mollis-commodo-mi%2C-sed-rutrum-ipsum-fringilla-at',
+            // },
+            {
+              title: 'Orci varius',
+              url: '#orci-varius',
+            },
+          ]"/>
+        </div>
+      </StickySection>
+    </template>
+
+    <!-- <BasicSection width="var(--main-width)" class="not-desktop">
+      TabletContentNav
+    </BasicSection> -->
+
+    <BasicSection width="var(--main-width)" class="component-border-vertical">
+      <Content v-if="!pendingPost && data?.post?.id" :content="data?.post?.content ?? ''" />
+      <ContentPlaceholder v-else />
+      <div class="component-border-top font-jost p-[30px]">
+        <PostAuthor v-if="!pendingPost && data?.post?.id" :profile="data.postAuthor" />
+        <PostAuthorPlaceholder v-else />
+      </div>
+    </BasicSection>
+  </SectionWrapper>
+
+  <SectionWrapper width="var(--desktop-main-content-width)">
+    <template #aside>
+      <StickySection width="340px">
+        <div class="m-[7px]">
+          <AdPlaceholder />
+        </div>
+      </StickySection>
+    </template>
+
+    <BasicSection width="var(--main-width)" class="component-border-vertical">
+      <Comments id="comments" :postId="data?.post?.id" />
+    </BasicSection>
+  </SectionWrapper>
 </template>
 
 <script lang="ts" setup>
@@ -53,57 +111,35 @@ const route = useRoute()
 const { data, pending: pendingPost } = useFetch<any>(`/api/posts/${route.meta.content_id}`)
 
 import { useGlobalStore } from '../store/global'
-const { getConnection } = useGlobalStore()
+const { getConnection, reconnect } = useGlobalStore()
 
 onMounted(() => {
   const connection = getConnection()
 
+  if (!connection || connection.readyState !== WebSocket.OPEN) {
+    reconnect()
+  }
+
   if (!connection) {
+    console.error('Connection failed!')
     return
   }
 
-  connection.onopen = () => {
-    connection.send(JSON.stringify({
+  connection.send(JSON.stringify({
       type: 'GET_POST_RATING',
       postId: route.meta.content_id,
-    }))
-  }
-
-  if (connection.readyState === WebSocket.OPEN) {
-    connection.send(JSON.stringify({
-      type: 'GET_POST_RATING',
-      postId: route.meta.content_id,
-    }))
-  }
+  }))
+  
 })
 
 const router = useRouter()
 
 watch(router.currentRoute, () => {
-  if (route.hash === '#comments') {
-    window.location.hash = ''
-
-    const comments = document.getElementById('comments')
-
-    if (!comments) {
-      return
-    }
-
-    const simpleContentWrapper = document.querySelector('.simplebar-content-wrapper')
-
-    if (!simpleContentWrapper) {
-      return
-    }
-
-    comments.scrollIntoView({
-      behavior: 'smooth',
-    })
-
-    simpleContentWrapper.scrollTo({
-      top: comments.offsetTop,
-      behavior: 'smooth',
-    })
+  if (!router.currentRoute.value.hash) {
+    return
   }
+
+  router.replace(router.currentRoute.value.path)
 }, {
   immediate: true,
 })